requirements.txt
matplotlib==3.5.1
pytest==6.2.5
PyYAML==6.0
spacy==3.2.1
torch==1.10.1
torchtext==0.11.1
tensorboard==2.7.0
